requirements.txt
accelerate==0.26.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
Pillow==10.2.0
Requests==2.31.0
rouge_score==0.1.2
torch==2.1.2+cu118
tqdm==4.65.0
transformers==4.32.1
